修改页面跳转前缀

+++ ruoyi-client/src/main/java/com/ruoyi/client/mapper/UserMapper.java
@@ -36,5 +36,5 @@ public interface UserMapper extends BaseMapper<User> {
      * @param userId 学生id
      * @return 学生信息
      */
-    UserInfoVO selectStudentInfo(Long userId);
+    UserInfoVO selectStudentInfo(Integer userId);
 }
